Payment Gateway and Leaderboard added

+++ controller/expense.js
@@ -2,30 +2,42 @@ const Expense = require('../models/expenses');
 
 const addexpense = (req, res) => {
     const { expenseamount, description, category } = req.body;
-    req.user.createExpense({ expenseamount, description, category }).then(expense => {
+
+    if(expenseamount == undefined || expenseamount.length === 0 ){
+        return res.status(400).json({success: false, message: 'Parameters missing'})
+    }
+    
+    Expense.create({ expenseamount, description, category, userId: req.user.id}).then(expense => {
         return res.status(201).json({expense, success: true } );
     }).catch(err => {
-        return res.status(403).json({success : false, error: err})
+        return res.status(500).json({success : false, error: err})
     })
 }
 
 const getexpenses = (req, res)=> {
-
-    req.user.getExpenses().then(expenses => {
+    
+    Expense.findAll({ where : { userId: req.user.id}}).then(expenses => {
         return res.status(200).json({expenses, success: true})
     })
     .catch(err => {
-        return res.status(402).json({ error: err, success: false})
+        console.log(err)
+        return res.status(500).json({ error: err, success: false})
     })
 }
 
 const deleteexpense = (req, res) => {
     const expenseid = req.params.expenseid;
-    Expense.destroy({where: { id: expenseid }}).then(() => {
-        return res.status(204).json({ success: true, message: "Deleted Successfuly"})
+    if(expenseid == undefined || expenseid.length === 0){
+        return res.status(400).json({success: false, })
+    }
+    Expense.destroy({where: { id: expenseid, userId: req.user.id }}).then((noofrows) => {
+        if(noofrows === 0){
+            return res.status(404).json({success: false, message: 'Expense doenst belong to the user'})
+        }
+        return res.status(200).json({ success: true, message: "Deleted Successfuly"})
     }).catch(err => {
         console.log(err);
-        return res.status(403).json({ success: true, message: "Failed"})
+        return res.status(500).json({ success: true, message: "Failed"})
     })
 }
